fix(api/applications): representations events refactor and typing (boas-1650)
