+++ apps/api/src/server/applications/application/representations/redact/redact.controller.js
@@ -1,7 +1,7 @@
 import { updateRepRedactRequestToRepository } from './redact.mapper.js';
 import { updateRedactedRepresentation } from './redact.service.js';
 import { EventType } from '@pins/event-client';
-import { sendRepresentationEventMessage } from '../representations.service.js';
+import { broadcastNsipRepresentationEvent } from '#infrastructure/event-broadcasters.js';
 import { getById } from '#repositories/representation.repository.js';
 
 /**
@@ -21,7 +21,7 @@ export const patchRepresentationRedact = async ({ params, body }, response) => {
 
 	// broadcast update event message
 	const representationFullDetails = await getById(Number(representationId));
-	await sendRepresentationEventMessage(representationFullDetails, EventType.Update);
+	await broadcastNsipRepresentationEvent(representationFullDetails, EventType.Update);
 
 	if (!representation) {
 		return response
