+++ apps/api/src/server/infrastructure/event-batch-broadcaster.js
@@ -1,14 +1,17 @@
-import { eventClient } from '#infrastructure/event-client.js';
+import { eventClient } from './event-client.js';
+
+const EVENT_BATCH_SIZE = 100;
 
 /**
+ * Send arrays of event messages in batches, to avoid 1mb throughput limit
  *
  * @param {string} topic
  * @param {any[]} events
- * @param {import('./event-type.js').EventType} eventType
+ * @param {import('@pins/event-client').EventType} eventType
  * @param {Object.<string,any>?} [additionalProperties={}]
  */
 export const batchSendEvents = async (topic, events, eventType, additionalProperties = {}) => {
-	const batchSize = 100;
+	const batchSize = EVENT_BATCH_SIZE;
 	// Send messages in chunks to avoid 1mb throughput limit
 	for (let i = 0; i < events.length; i += batchSize) {
 		const batch = events.slice(i, i + batchSize);
