+++ apps/api/src/server/infrastructure/event-broadcasters.js
@@ -1,7 +1,16 @@
 import logger from '#utils/logger.js';
 import { eventClient } from './event-client.js';
-import { FOLDER, NSIP_PROJECT, SERVICE_USER, NSIP_S51_ADVICE } from './topics.js';
-import { verifyNotTraining } from '../applications/application/application.validators.js';
+import {
+	FOLDER,
+	NSIP_PROJECT,
+	SERVICE_USER,
+	NSIP_S51_ADVICE,
+	NSIP_REPRESENTATION
+} from './topics.js';
+import {
+	isTrainingCase,
+	verifyNotTraining
+} from '../applications/application/application.validators.js';
 import { EventType } from '@pins/event-client';
 import { getAllByCaseId } from '#repositories/folder.repository.js';
 import { filterAsync } from '#utils/async.js';
@@ -11,11 +20,18 @@ import { buildServiceUserPayload } from '#infrastructure/payload-builders/applic
 import { buildFoldersPayload } from '#infrastructure/payload-builders/folder.js';
 import { buildNsipS51AdvicePayload } from '#infrastructure/payload-builders/nsip-s51-advice.js';
 import { verifyNotTrainingS51 } from '../applications/s51advice/s51-advice.validators.js';
+import {
+	buildNsipRepresentationPayload,
+	buildNsipRepresentationPayloadForPublish,
+	buildRepresentationServiceUserPayload
+} from './payload-builders/nsip-representation.js';
+import { batchSendEvents } from './event-batch-broadcaster.js';
 
 const applicant = 'Applicant';
 
 /**
  * @typedef {import('@pins/applications.api').Schema.S51Advice} S51Advice
+ * @typedef {import('@prisma/client').Prisma.RepresentationGetPayload<{include: {case: true, user: true, represented: true, representative: true, attachments: true, representationActions: true} }>} RepresentationWithFullDetails
  */
 
 /**
@@ -87,5 +103,62 @@ export const broadcastNsipS51AdviceEvent = async (s51Advice, eventType) => {
 
 	if (eventPayloads.length) {
 		await eventClient.sendEvents(NSIP_S51_ADVICE, eventPayloads, eventType);
+	}
+};
+
+/**
+ * Broadcast a create / update event message to Service Bus, for a representation, and any service users (reps contact or agent)
+ *
+ * @param {RepresentationWithFullDetails} representation
+ * @param {EventType} eventType
+ * @returns
+ */
+export const broadcastNsipRepresentationEvent = async (
+	representation,
+	eventType = EventType.Update
+) => {
+	// dont send events for TRAINING cases
+	if (!isTrainingCase(representation.case.caseRef)) {
+		const nsipRepresentationPayload = buildNsipRepresentationPayload(representation);
+		const serviceUsersPayload = buildRepresentationServiceUserPayload(representation);
+
+		await eventClient.sendEvents(NSIP_REPRESENTATION, [nsipRepresentationPayload], eventType);
+
+		// and service users
+		await eventClient.sendEvents(SERVICE_USER, serviceUsersPayload, eventType, {
+			entityType: 'RepresentationContact'
+		});
+	}
+};
+
+/**
+ *
+ * @param {RepresentationWithFullDetails[]} representations
+ * @param {EventType} eventType
+ * @param {number} caseId
+ */
+export const broadcastNsipRepresentationPublishEventBatch = async (
+	representations,
+	eventType = EventType.Publish,
+	caseId
+) => {
+	// pull the ids from the reps array
+	const representationIds = representations.map((rep) => rep.id);
+	// TODO: DJW check this
+
+	// dont send events for TRAINING cases
+	const nsipRepresentationsPayload = representations.map(buildNsipRepresentationPayloadForPublish);
+	const serviceUsersPayload = representations.flatMap(buildRepresentationServiceUserPayload);
+
+	try {
+		await verifyNotTraining(caseId);
+
+		await batchSendEvents(NSIP_REPRESENTATION, nsipRepresentationsPayload, eventType);
+
+		await batchSendEvents(SERVICE_USER, serviceUsersPayload, eventType, {
+			entityType: 'RepresentationContact'
+		});
+	} catch (/** @type {*} */ err) {
+		logger.info(`Blocked sending event for representations: ${representationIds}`, err.message);
 	}
 };
